+++ datacommons/__init__.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 
 # Data Commons SPARQL query support
-from datacommons.query import Query
+from datacommons.query import query
 
 # Data Commons Python Client API
 from datacommons.core import get_property_labels, get_property_values, get_triples
 from datacommons.places import get_places_in
-from datacommons.populations import get_populations, get_observations
+from datacommons.populations import get_populations, get_observations, get_pop_obs, get_place_obs
 
 # Other utilities
 from .utils import set_api_key, clean_frame, flatten_frame